refactor!: fix duplication of component type basecomponentprops (#420)
