--- src/theme/common/entities.ts
@@ -1,7 +1,0 @@
-export interface BaseComponentProps {
-  className?: string;
-}
-
-export interface TrackingComponentProps {
-  trackingId?: string;
-}
